게시물 수정, 삭제 권한체크, 로그인체크. jsp include 사용법

+++ src/main/webapp/jsp/article/detail.jsp
@@ -7,11 +7,8 @@
 
 <%
 	Map<String, Object> articleRow = (Map) request.getAttribute("articleRow");
-%>
-<%
 	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
 %>
-
 
 <html>
 <head>
@@ -19,10 +16,6 @@
 <title>게시물 상세보기</title>
 </head>
 <body>
-
-
-
-
 	<h1><%= (int) articleRow.get("id") %>번 게시물</h1>
 	
 	<div>번호 : <%= (int) articleRow.get("id") %></div>
@@ -31,17 +24,16 @@
 	<div>제목 : <%= (String) articleRow.get("title") %></div>
 	<div>내용 : <%= (String) articleRow.get("body") %></div>
 	<div>
-	
-		<a href="list">목록</a>	
-	<%
-	if (loginedMemberId != -1) {
-	%>
-	
-		<a href="modify?id=<%= (int) articleRow.get("id") %>">수정</a>
-		<a href="doDelete?id=<%= (int) articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
-	<%
-	}
-	%>
+		<a href="list">목록</a>
+		<%
+		if (loginedMemberId != -1 && (int) articleRow.get("memberId") == loginedMemberId) {
+		%>
+			<a href="modify?id=<%= (int) articleRow.get("id") %>">수정</a>
+			<a href="doDelete?id=<%= (int) articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		<%
+		}
+		%>
 	</div>
+	
 </body>
 </html>
